Readjust project; Tested simple jsp and constoller
