--- src/main/java/com/rayedchan/hackruspring2015/Language.java
@@ -1,25 +1,0 @@
-package com.rayedchan.hackruspring2015;
-
-/**
- * Represents a single language
- * @author rayedchan
- */
-public class Language 
-{
-    private String language;
-    
-    public String getLanguage()
-    {
-        return this.language;
-    }
-    public void setLanguage(String language)
-    {
-        this.language = language;
-    }
-    
-    @Override
-    public String toString()
-    {
-        return String.format("%s", this.language);
-    }
-}
