+++ src/main/java/edu/rutgers/hackruspring2015/controllers/TranslateController.java
@@ -1,0 +1,26 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package edu.rutgers.hackruspring2015.controllers;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.servlet.ModelAndView;
+
+/**
+ *
+ * @author rayedchan
+ */
+@Controller
+public class TranslateController 
+{
+   @RequestMapping("/translate")
+   public ModelAndView helloWorld() 
+   {
+     String message = "<br><div style='text-align:center;'>"
+				+ "<h3>********** Hello World, Spring MVC Tutorial</h3>This message is coming from CrunchifyHelloWorld.java **********</div><br><br>";
+		return new ModelAndView("translate", "message", message);
+   }
+}
