+++ src/main/webapp/WEB-INF/jsp/translate.jsp
@@ -1,0 +1,17 @@
+<%-- 
+    Document   : translate
+    Created on : Apr 27, 2015, 9:20:46 PM
+    Author     : rayedchan
+--%>
+
+<%@page contentType="text/html" pageEncoding="UTF-8"%>
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+        <title>Google Translate API</title>
+    </head>
+    <body>
+        <h1>${message}</h1>
+    </body>
+</html>
